feat: add ability to modify max-jobs at runtime via stdin
This change replaces the sleep call with a getch+timeout call.  If no
input is given, the application behaves as before.  Because of the
timeout, it is effectively a sleep call.  If input is given, it is
parsed for character based commands.

During getch, SIGINT is disabled.  Ctrl-C will generate a character
instead.  If we see this character, we respond to it the same as a
SIGINT.

If a '=' character is seen, then we prompt for a new max-jobs value.