feat!: change output directory relative path expansion behavior
Previous to this change, the output directory was relative to the
script.  The intent was for the output directory command-line option to
be relative to PWD while the output directory script option would be
relative to the script.  Handling them differently requires too much
complexity for little to no benefit.

It makes more sense for output directory to be relative to PWD.  If
scripts want to make relative to the script, then they can use
`File.join(__dir__, path_relative_to_script)`.

BREAKING CHANGE: Changes relative path handling for
`jobrnr_options.output_directory` in user scripts. Was relative to
script.  Now relative to PWD.