fix: child processes not receiving signals (e.g. SIGINT)
Previous to this change, jobs were spawned using spawn(1).  This
invocation of spawn() uses a shell (e.g. sh -c job).  When we send a
signal to the spawned PID, we were actually sending a signal to the sh
process and the signal wasn't propagating to the job process
(grandchild) for some reason.

Fix by using the spawn(3) invocation.  This form does not spawn a
subshell.  Because a subshell is no longer being spawned, we lose some
of the functionality that the shell provided: quoted arg handling,
environment variable expansion, and others(?).

Quoted arg handling is a must, so now we have to do it ourselves.
Luckily Ruby provides the handy shellwords module that implements this.

Environment variable expansion was never advertised as a feature so I'm
fine with losing this functionality.